+++ credativ_pg_migrator/database_connector.py
@@ -1,16 +1,16 @@
 # credativ-pg-migrator
 # Copyright (C) 2025 credativ GmbH
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
-# 
+#
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -66,6 +66,20 @@ class DatabaseConnector(ABC):
             'comment': table_comment
             }
         }
+        """
+        pass
+
+    @abstractmethod
+    def get_table_description(self, settings) -> dict:
+        """
+        settings - dictionary with the following keys
+            - table_schema: str,
+            - table_name: str,
+        Fetch a description of the table returned by the source database.
+        Content depends on the database type.
+        Added for better observability of the migration process.
+        Returns a simple dictionary:
+            - 'table_description': description of the table from the source database
         """
         pass